doc capture: validate that artifact name is unique
